extract chain handlers to separate files
